--- tests/kernel/systhreads/README
@@ -1,12 +1,0 @@
-$ make run
-
-Running test suite test_systhreads
-tc_start() - test_systhreads_main
-===================================================================
-PASS - test_systhreads_main.
-tc_start() - test_systhreads_idle
-===================================================================
-PASS - test_systhreads_idle.
-===================================================================
-PROJECT EXECUTION SUCCESSFUL
-
