--- tests/kernel/systhreads/prj.conf
@@ -1,4 +1,0 @@
-CONFIG_ZTEST=y
-#CONFIG_SYS_POWER_MANAGEMENT=y
-#CONFIG_SYS_POWER_LOW_POWER_STATE=y
-CONFIG_TICKLESS_IDLE=y
